Ver. 5.0 + Mobile 3
I changed the styles, cut back the replacement of the background when choosing a different "ColorRing", updated the layout and added a little responsive.

You can also notice the presence of a folder for Mobile. The page looks tolerable on smartphones, but the arrangement of buttons, styles (more related to Android) needs to be improved.

From grandiose plans:
- Add the ability to save reports about incorrect results (I think this is a database in the form of a table)
- Improve ergonomic design to reduce the possibility of accidentally clicking on elements
- (Less grandiose) Add numbering to each displayed result from the first + 1. Orient the order of results: last, previous, first (top to bottom).